backtests/requirements.txt
polars>=0.20.19
python-dateutil>=2.9.0
